+++ CHANGES.md
@@ -47,7 +47,9 @@ Gateway:
 Authorization:
 
 -   Made integration tests for authorisation run automatically as part of CI.
--   Added ability to set per-record authorization policies in the registry
+-   Added ability to set per-record authorization policies in the registry (for getting a single record)
+-   Added ability to set per-record authorization policies in the registry (for getting multiple records)
+-   Added ability to use OPA policies that use data types other than strings in the registry
 
 Others: